+++ wasmlisp/src/lisp.rs
@@ -2,16 +2,6 @@
    Rust study program.
    This is prototype program mini scheme subset what porting from go-scheme.
 
-   1. howto test)
-      wasm-pack test --headless --chrome -- --lib
-       or
-      wasm-pack test --headless --firefox -- --lib
-
-   2. build & run)
-      wasm-pack build
-      npm install
-      npm run lisp
-
    hidekuno@gmail.com
 */
 extern crate elisp;
@@ -34,6 +24,11 @@ use lisp::Environment;
 use lisp::Expression;
 use lisp::ErrCode;
 use lisp::Error;
+use crate::draw::create_draw_string;
+use crate::draw::create_draw_line;
+use crate::draw::create_draw_arc;
+use crate::draw::create_draw_image;
+use crate::draw::Graphics;
 
 use std::cell::RefCell;
 use std::io::Cursor;
@@ -105,22 +100,9 @@ pub fn start() -> Result<(), JsValue> {
     Ok(())
 }
 //--------------------------------------------------------
-// Graphics (ex. background color)
-//--------------------------------------------------------
-struct Graphics {
-    bg: Option<JsValue>,
-}
-impl Graphics {
-    fn new() -> Graphics {
-        Graphics {
-            bg: None,
-        }
-    }
-}
-//--------------------------------------------------------
 // lisp functions
 //--------------------------------------------------------
-fn build_lisp_function(env: &Environment, document: &web_sys::Document) {
+pub fn build_lisp_function(env: &Environment, document: &web_sys::Document) {
     let canvas = document
         .get_element_by_id("drawingarea")
         .unwrap()
@@ -160,8 +142,9 @@ fn build_lisp_function(env: &Environment, document: &web_sys::Document) {
     // Draw Line
     // ex. (draw-line 10.0 10.0 100.0 100.0)
     //--------------------------------------------------------
-    let ctx = context.clone();
-    ctx.set_line_width(LINE_WIDTH);
+    context.set_line_width(LINE_WIDTH);
+
+    let draw_line = create_draw_line(&context);
     env.add_builtin_ext_func("draw-line", move |exp, env| {
         const N: usize = 4;
         if exp.len() != (N + 1) {
@@ -178,11 +161,7 @@ fn build_lisp_function(env: &Environment, document: &web_sys::Document) {
                 }
             }
         }
-        ctx.begin_path();
-        ctx.move_to(loc[0], loc[1]);
-        ctx.line_to(loc[2], loc[3]);
-        ctx.close_path();
-        ctx.stroke();
+        draw_line(loc[0], loc[1],loc[2], loc[3]);
         Ok(Expression::Nil())
     });
     //--------------------------------------------------------
@@ -218,8 +197,7 @@ fn build_lisp_function(env: &Environment, document: &web_sys::Document) {
     //--------------------------------------------------------
     // ex. (draw-image "roger" 0.0 0.0 180.0 0.0 0.0 180.0)
     //--------------------------------------------------------
-    let doc = document.clone();
-    let ctx = context.clone();
+    let draw_image = create_draw_image(&context,&document);
     env.add_builtin_ext_func("draw-image", move |exp, env| {
         if exp.len() != 8 {
             return Err(create_error!(ErrCode::E1007));
@@ -242,30 +220,7 @@ fn build_lisp_function(env: &Environment, document: &web_sys::Document) {
                 return Err(create_error!(ErrCode::E1007));
             }
         }
-        let img = match doc.get_element_by_id(&symbol) {
-            Some(e) => e.dyn_into::<HtmlImageElement>().unwrap(),
-            None => return Err(create_error!(ErrCode::E1008)),
-        };
-        let w = img.width() as f64;
-        let h = img.height() as f64;
-        if let Err(_) = ctx.set_transform(
-            ctm[2] / w,
-            ctm[3] / h,
-            ctm[4] / w,
-            ctm[5] / h,
-            ctm[0],
-            ctm[1],
-        ) {
-            return Err(create_error!(ErrCode::E9999));
-        }
-
-        // https://rustwasm.github.io/wasm-bindgen/api/web_sys/
-        if let Err(_) = ctx.draw_image_with_html_image_element(&img, 0.0, 0.0) {
-            return Err(create_error!(ErrCode::E9999));
-        }
-        if let Err(_) = ctx.set_transform(1.0, 0.0, 0.0, 1.0, 0.0, 0.0) {
-            return Err(create_error!(ErrCode::E9999));
-        }
+        draw_image(ctm[0],ctm[1],ctm[2],ctm[3],ctm[4],ctm[5],&symbol)?;
         Ok(Expression::Nil())
     });
     //--------------------------------------------------------
@@ -332,24 +287,6 @@ fn build_lisp_function(env: &Environment, document: &web_sys::Document) {
         let (_, h) = image_size(exp, env, &doc)?;
         Ok(Expression::Float(h))
     });
-    fn image_size(
-        exp: &[Expression],
-        env: &Environment,
-        doc: &Document,
-    ) -> Result<(f64, f64), Error> {
-        if exp.len() != 2 {
-            return Err(create_error!(ErrCode::E1007));
-        }
-        let symbol = match eval(&exp[1], env)? {
-            Expression::String(s) => s,
-            _ => return Err(create_error!(ErrCode::E1015)),
-        };
-        let img = match doc.get_element_by_id(&symbol) {
-            Some(e) => e.dyn_into::<HtmlImageElement>().unwrap(),
-            None => return Err(create_error!(ErrCode::E9999)),
-        };
-        Ok((img.width() as f64, img.height() as f64))
-    }
     //--------------------------------------------------------
     // ex. (load-url "sicp/segments-fish.scm")
     //--------------------------------------------------------
@@ -501,7 +438,7 @@ fn build_lisp_function(env: &Environment, document: &web_sys::Document) {
     // Draw arc
     // ex. (draw-arc 75.0 75.0 50.0 0.0)
     //--------------------------------------------------------
-    let ctx = context.clone();
+    let draw_arc =  create_draw_arc(&context);
     env.add_builtin_ext_func("draw-arc", move |exp, env| {
         if exp.len() != 5 {
             return Err(create_error!(ErrCode::E1007));
@@ -514,12 +451,9 @@ fn build_lisp_function(env: &Environment, document: &web_sys::Document) {
                 _ => return Err(create_error!(ErrCode::E1003)),
             };
         }
-        ctx.begin_path();
-        ctx.arc(prm[0], prm[1], prm[2], prm[3], std::f64::consts::PI * 2.0).unwrap();
-        ctx.stroke();
+        draw_arc(prm[0], prm[1], prm[2], prm[3]);
         Ok(Expression::Nil())
     });
-
     //--------------------------------------------------------
     // Draw string
     // ex. (draw-string "hello,world" 0.0 10.0)
@@ -577,17 +511,25 @@ fn build_lisp_function(env: &Environment, document: &web_sys::Document) {
     });
 }
 // ----------------------------------------------------------------
-// create new cairo from imagetable, and draw string
+// image size
 // ----------------------------------------------------------------
-pub fn create_draw_string(context: &CanvasRenderingContext2d)
-                          -> Box<dyn Fn(String, f64, f64, String) + 'static> {
-
-    let ctx = context.clone();
-    let draw_string = move |s: String, x, y, f:String| {
-        ctx.set_font(&f);
-        ctx.fill_text(&s, x, y).unwrap();
+fn image_size(
+    exp: &[Expression],
+    env: &Environment,
+    doc: &Document,
+) -> Result<(f64, f64), Error> {
+    if exp.len() != 2 {
+        return Err(create_error!(ErrCode::E1007));
+    }
+    let symbol = match eval(&exp[1], env)? {
+        Expression::String(s) => s,
+        _ => return Err(create_error!(ErrCode::E1015)),
     };
-    Box::new(draw_string)
+    let img = match doc.get_element_by_id(&symbol) {
+        Some(e) => e.dyn_into::<HtmlImageElement>().unwrap(),
+        None => return Err(create_error!(ErrCode::E9999)),
+    };
+    Ok((img.width() as f64, img.height() as f64))
 }
 // ----------------------------------------------------------------
 // load scheme program
@@ -616,481 +558,4 @@ async fn get_program_file(scm: String) -> Result<JsValue, JsValue> {
     let text = JsFuture::from(resp.text()?).await?;
     console_log!("http get complete: {}", scm);
     Ok(text)
-}
-#[cfg(test)]
-mod tests {
-    use super::*;
-    extern crate wasm_bindgen_test;
-    use wasm_bindgen_test::*;
-
-    const IMG_URL: &'static str =
-        "https://github.com/hidekuno/picture-language/blob/master/sicp/sicp.png?raw=true";
-
-    const SD_URL: &'static str =
-        "https://coverartarchive.org/release-group/9b1acd78-3d19-37bb-8ca0-5816d44da439/front-250.jpg";
-
-    const RV_URL: &'static str =
-        "https://coverartarchive.org/release-group/72d15666-99a7-321e-b1f3-a3f8c09dff9f/front-250.jpg";
-
-    const PS_URL: &'static str =
-        "https://coverartarchive.org/release-group/fdd96703-7b21-365e-bdea-38029fbeb84e/front-250.jpg";
-
-    const CANVAS_WIDTH: u32 = 720;
-    const CANVAS_HEIGHT: u32 = 560;
-
-    wasm_bindgen_test_configure!(run_in_browser);
-
-    fn do_lisp_env(program: &str, env: &Environment) -> String {
-        match do_core_logic(&program.into(), env) {
-            Ok(v) => v.to_string(),
-            Err(e) => e.get_code(),
-        }
-    }
-    fn create_document() -> Document {
-        let document = web_sys::window().unwrap().document().unwrap();
-        document.create_element("body").unwrap();
-
-        let canvas = document
-            .create_element("canvas")
-            .unwrap()
-            .dyn_into::<HtmlCanvasElement>()
-            .unwrap();
-
-        canvas.set_id("drawingarea");
-        canvas.set_width(CANVAS_WIDTH);
-        canvas.set_height(CANVAS_HEIGHT);
-        document.body().unwrap().append_child(&canvas).unwrap();
-        document
-    }
-    #[wasm_bindgen_test]
-    fn draw_clear() {
-        let document = create_document();
-        let env = Environment::new();
-        build_lisp_function(&env, &document);
-        assert_eq!(do_lisp_env("(draw-clear)", &env), "nil");
-    }
-    #[wasm_bindgen_test]
-    fn draw_line() {
-        let document = create_document();
-        let env = Environment::new();
-        build_lisp_function(&env, &document);
-        assert_eq!(
-            do_lisp_env("(draw-line 10.0 10.0 100.0 100.0)", &env),
-            "nil"
-        );
-    }
-    #[wasm_bindgen_test]
-    fn gtk_major_version() {
-        let document = create_document();
-        let env = Environment::new();
-        build_lisp_function(&env, &document);
-        assert_eq!(do_lisp_env("(gtk-major-version)", &env), "-1");
-    }
-    #[wasm_bindgen_test]
-    fn screen_width() {
-        let document = create_document();
-        let env = Environment::new();
-        build_lisp_function(&env, &document);
-        assert_eq!(do_lisp_env("(screen-width)", &env), CANVAS_WIDTH.to_string());
-    }
-    #[wasm_bindgen_test]
-    fn screen_height() {
-        let document = create_document();
-        let env = Environment::new();
-        build_lisp_function(&env, &document);
-        assert_eq!(do_lisp_env("(screen-height)", &env), CANVAS_HEIGHT.to_string());
-    }
-    #[wasm_bindgen_test]
-    fn load_image() {
-        let document = create_document();
-        let env = Environment::new();
-        build_lisp_function(&env, &document);
-        assert_eq!(
-            do_lisp_env(
-                format!("(load-image \"roger\" \"{}\")", IMG_URL).as_str(),
-                &env
-            ),
-            "nil"
-        );
-    }
-    #[wasm_bindgen_test]
-    fn draw_image() {
-        let document = create_document();
-        let env = Environment::new();
-        build_lisp_function(&env, &document);
-        do_lisp_env(
-            format!("(load-image \"roger\" \"{}\")", SD_URL).as_str(),
-            &env,
-        );
-        assert_eq!(
-            do_lisp_env("(draw-image \"roger\" 0.0 0.0 1.0 0.0 0.0 1.0)", &env),
-            "nil"
-        );
-    }
-    #[wasm_bindgen_test]
-    fn image_width() {
-        let document = create_document();
-        let env = Environment::new();
-        build_lisp_function(&env, &document);
-        do_lisp_env(
-            format!("(load-image \"roger\" \"{}\")", RV_URL).as_str(),
-            &env,
-        );
-        // NG because It's Asynchronous processing
-        assert_eq!(do_lisp_env("(image-width \"roger\")", &env), "0");
-    }
-    #[wasm_bindgen_test]
-    fn image_height() {
-        let document = create_document();
-        let env = Environment::new();
-        build_lisp_function(&env, &document);
-        do_lisp_env(
-            format!("(load-image \"roger\" \"{}\")", PS_URL).as_str(),
-            &env,
-        );
-        // NG because It's Asynchronous processing
-        assert_eq!(do_lisp_env("(image-height \"roger\")", &env), "0");
-    }
-    #[wasm_bindgen_test]
-    fn load_url() {
-        let document = create_document();
-        let env = Environment::new();
-        build_lisp_function(&env, &document);
-        assert_eq!(
-            do_lisp_env("(load-url \"sicp/abstract-data.scm\")", &env),
-            "nil"
-        );
-        // NG because It's Asynchronous processing
-        // left: `"E1008"`,
-        // right: `"Function"`', src/lisp.rs:493:9
-        // assert_eq!(do_lisp_env("make-frame", &env), "Function");
-    }
-    #[wasm_bindgen_test]
-    fn add_timeout() {
-        let document = create_document();
-        let env = Environment::new();
-        build_lisp_function(&env, &document);
-        assert_eq!(do_lisp_env("(add-timeout (+ 1 1) 10)", &env), "nil");
-    }
-    #[wasm_bindgen_test]
-    fn set_foreground() {
-        let document = create_document();
-        let env = Environment::new();
-        build_lisp_function(&env, &document);
-        assert_eq!(
-            do_lisp_env("(set-foreground \"blue\")", &env),
-            "nil"
-        );
-    }
-    #[wasm_bindgen_test]
-    fn set_background() {
-        let document = create_document();
-        let env = Environment::new();
-        build_lisp_function(&env, &document);
-        assert_eq!(
-            do_lisp_env("(set-background \"black\")", &env),
-            "nil"
-        );
-    }
-    #[wasm_bindgen_test]
-    fn set_line_width() {
-        let document = create_document();
-        let env = Environment::new();
-        build_lisp_function(&env, &document);
-        assert_eq!(
-            do_lisp_env("(set-line-width 1.0)", &env),
-            "nil"
-        );
-    }
-    #[wasm_bindgen_test]
-    fn draw_arc() {
-        let document = create_document();
-        let env = Environment::new();
-        build_lisp_function(&env, &document);
-        assert_eq!(
-            do_lisp_env("(draw-arc 75.0 75.0 50.0 0.0)", &env),
-            "nil"
-        );
-    }
-    #[wasm_bindgen_test]
-    fn draw_string() {
-        let document = create_document();
-        let env = Environment::new();
-        build_lisp_function(&env, &document);
-        assert_eq!(
-            do_lisp_env("(draw-string \"Hello,World\" 20.0 20.0)", &env),
-            "nil"
-        );
-        assert_eq!(
-            do_lisp_env("(draw-string \"Hello,World\" 20.0 20.0 \"italic bold 20px sans-serif\")", &env),
-            "nil"
-        );
-    }
-    #[wasm_bindgen_test]
-    fn draw_eval() {
-        let document = create_document();
-        let env = Environment::new();
-        build_lisp_function(&env, &document);
-        assert_eq!(do_lisp_env("(draw-eval (iota 20))", &env), "nil");
-    }
-}
-#[cfg(test)]
-mod error_tests {
-    use super::*;
-    extern crate wasm_bindgen_test;
-    use wasm_bindgen_test::*;
-
-    const IMG_URL: &'static str =
-        "https://github.com/hidekuno/picture-language/blob/master/sicp/sicp.png?raw=true";
-
-    const CANVAS_WIDTH: u32 = 720;
-    const CANVAS_HEIGHT: u32 = 560;
-
-    wasm_bindgen_test_configure!(run_in_browser);
-    fn do_lisp_env(program: &str, env: &Environment) -> String {
-        match do_core_logic(&program.into(), env) {
-            Ok(v) => v.to_string(),
-            Err(e) => e.get_code(),
-        }
-    }
-    fn create_document() -> Document {
-        let document = web_sys::window().unwrap().document().unwrap();
-        document.create_element("body").unwrap();
-
-        let canvas = document
-            .create_element("canvas")
-            .unwrap()
-            .dyn_into::<HtmlCanvasElement>()
-            .unwrap();
-
-        canvas.set_id("drawingarea");
-        canvas.set_width(CANVAS_WIDTH);
-        canvas.set_height(CANVAS_HEIGHT);
-        document.body().unwrap().append_child(&canvas).unwrap();
-        document
-    }
-    #[wasm_bindgen_test]
-    fn draw_clear() {
-        let document = create_document();
-        let env = Environment::new();
-        build_lisp_function(&env, &document);
-        assert_eq!(do_lisp_env("(draw-clear 1)", &env), "E1007");
-    }
-    #[wasm_bindgen_test]
-    fn draw_line() {
-        let document = create_document();
-        let env = Environment::new();
-        build_lisp_function(&env, &document);
-        assert_eq!(do_lisp_env("(draw-line)", &env), "E1007");
-        assert_eq!(do_lisp_env("(draw-line 0.0 1.0 2.0 3)", &env), "E1003");
-        assert_eq!(do_lisp_env("(draw-line a b 2.0 3)", &env), "E1008");
-        assert_eq!(do_lisp_env("(draw-line 0.0 1.0 2.0 3)", &env), "E1003");
-        assert_eq!(do_lisp_env("(draw-line a b 2.0 3)", &env), "E1008");
-    }
-    #[wasm_bindgen_test]
-    fn gtk_major_version() {
-        let document = create_document();
-        let env = Environment::new();
-        build_lisp_function(&env, &document);
-        assert_eq!(do_lisp_env("(gtk-major-version 1)", &env), "E1007");
-    }
-    #[wasm_bindgen_test]
-    fn screen_width() {
-        let document = create_document();
-        let env = Environment::new();
-        build_lisp_function(&env, &document);
-        assert_eq!(do_lisp_env("(screen-width 1)", &env), "E1007");
-    }
-    #[wasm_bindgen_test]
-    fn screen_height() {
-        let document = create_document();
-        let env = Environment::new();
-        build_lisp_function(&env, &document);
-        assert_eq!(do_lisp_env("(screen-height 1)", &env), "E1007");
-    }
-    #[wasm_bindgen_test]
-    fn load_image() {
-        let document = create_document();
-        let env = Environment::new();
-        build_lisp_function(&env, &document);
-        assert_eq!(do_lisp_env("(load-image)", &env), "E1007");
-        assert_eq!(do_lisp_env("(load-image  \"sample\")", &env), "E1007");
-        assert_eq!(
-            do_lisp_env("(load-image  \"sample\" 10 20 30)", &env),
-            "E1007"
-        );
-        assert_eq!(
-            do_lisp_env(format!("(load-image 10 \"{}\")", IMG_URL).as_str(), &env),
-            "E1015"
-        );
-        assert_eq!(do_lisp_env("(load-image \"sample\" 10)", &env), "E1015");
-    }
-    #[wasm_bindgen_test]
-    fn draw_image() {
-        let document = create_document();
-        let env = Environment::new();
-        build_lisp_function(&env, &document);
-        do_lisp_env(
-            format!("(load-image \"sample\" \"{}\")", IMG_URL).as_str(),
-            &env,
-        );
-        assert_eq!(do_lisp_env("(draw-image)", &env), "E1007");
-        assert_eq!(do_lisp_env("(draw-image 10)", &env), "E1007");
-        assert_eq!(
-            do_lisp_env("(draw-image \"sample\" 1 2 3 10)", &env),
-            "E1007"
-        );
-        assert_eq!(
-            do_lisp_env("(draw-image 10 0.0 0.0 1.0 0.0 0.0 1.0)", &env),
-            "E1015"
-        );
-        assert_eq!(
-            do_lisp_env("(draw-image \"sample1\" 0.0 0.0 1.0 0.0 0.0 1.0)", &env),
-            "E1008"
-        );
-        assert_eq!(
-            do_lisp_env("(draw-image \"sample\" 0.0 0.0 1.0 1.0 1.0 10)", &env),
-            "E1003"
-        );
-        assert_eq!(
-            do_lisp_env("(draw-image \"sample\" a 0.0 1.0 1.0 1.0 10)", &env),
-            "E1008"
-        );
-    }
-    #[wasm_bindgen_test]
-    fn image_width() {
-        let document = create_document();
-        let env = Environment::new();
-        build_lisp_function(&env, &document);
-        do_lisp_env(
-            format!("(load-image \"sample\" \"{}\")", IMG_URL).as_str(),
-            &env,
-        );
-        assert_eq!(do_lisp_env("(image-width)", &env), "E1007");
-        assert_eq!(
-            do_lisp_env("(image-width \"sample\" \"sample1\")", &env),
-            "E1007"
-        );
-        assert_eq!(do_lisp_env("(image-width 10)", &env), "E1015");
-        assert_eq!(do_lisp_env("(image-width a)", &env), "E1008");
-    }
-    #[wasm_bindgen_test]
-    fn image_height() {
-        let document = create_document();
-        let env = Environment::new();
-        build_lisp_function(&env, &document);
-        do_lisp_env(
-            format!("(load-image \"sample\" \"{}\")", IMG_URL).as_str(),
-            &env,
-        );
-        assert_eq!(do_lisp_env("(image-height)", &env), "E1007");
-        assert_eq!(
-            do_lisp_env("(image-height \"sample\" \"sample1\")", &env),
-            "E1007"
-        );
-        assert_eq!(do_lisp_env("(image-height 10)", &env), "E1015");
-        assert_eq!(do_lisp_env("(image-height a)", &env), "E1008");
-    }
-    #[wasm_bindgen_test]
-    fn load_url() {
-        let document = create_document();
-        let env = Environment::new();
-        build_lisp_function(&env, &document);
-        assert_eq!(do_lisp_env("(load-url)", &env), "E1007");
-        assert_eq!(
-            do_lisp_env("(load-url \"sicp/abstract-data.scm\" 10 12)", &env),
-            "E1007"
-        );
-        assert_eq!(do_lisp_env("(load-url 10)", &env), "E1015");
-        assert_eq!(do_lisp_env("(load-url a)", &env), "E1008");
-    }
-    #[wasm_bindgen_test]
-    fn add_timeout() {
-        let document = create_document();
-        let env = Environment::new();
-        build_lisp_function(&env, &document);
-        assert_eq!(do_lisp_env("(add-timeout)", &env), "E1007");
-        assert_eq!(do_lisp_env("(add-timeout (+ 1 1) 10 10)", &env), "E1007");
-        assert_eq!(do_lisp_env("(add-timeout (+ 1 1) #t)", &env), "E1002");
-        assert_eq!(do_lisp_env("(add-timeout (+ 1 1) 0)", &env), "E1021");
-    }
-        #[wasm_bindgen_test]
-    fn set_foreground() {
-        let document = create_document();
-        let env = Environment::new();
-        build_lisp_function(&env, &document);
-
-        assert_eq!(do_lisp_env("(set-foreground)", &env),"E1007");
-        assert_eq!(do_lisp_env("(set-foreground \"black\" \"black\")", &env),"E1007");
-        assert_eq!(do_lisp_env("(set-foreground #t)", &env),"E1015");
-    }
-    #[wasm_bindgen_test]
-    fn set_background() {
-        let document = create_document();
-        let env = Environment::new();
-        build_lisp_function(&env, &document);
-
-        assert_eq!(do_lisp_env("(set-background)", &env),"E1007");
-        assert_eq!(do_lisp_env("(set-background \"black\" \"black\")", &env),"E1007");
-        assert_eq!(do_lisp_env("(set-background #t)", &env),"E1015");
-    }
-    #[wasm_bindgen_test]
-    fn set_line_width() {
-        let document = create_document();
-        let env = Environment::new();
-        build_lisp_function(&env, &document);
-
-        assert_eq!(do_lisp_env("(set-line-width)", &env),"E1007");
-        assert_eq!(do_lisp_env("(set-line-width 1.0 1.0)", &env),"E1007");
-        assert_eq!(do_lisp_env("(set-line-width #t)", &env),"E1003");
-    }
-    #[wasm_bindgen_test]
-    fn draw_arc() {
-        let document = create_document();
-        let env = Environment::new();
-        build_lisp_function(&env, &document);
-
-        assert_eq!(do_lisp_env("(draw-arc)", &env),"E1007");
-        assert_eq!(do_lisp_env("(draw-arc 0.27 0.65 0.02 0.0 1.0)", &env),"E1007");
-        assert_eq!(do_lisp_env("(draw-arc #t 0.65 0.02 0.0)", &env),"E1003");
-        assert_eq!(do_lisp_env("(draw-arc 0.27 0.65 0.02 #t)", &env),"E1003");
-        assert_eq!(do_lisp_env("(draw-arc 0.27 0.65 #t 0.0)", &env),"E1003");
-    }
-    #[wasm_bindgen_test]
-    fn draw_string() {
-        let document = create_document();
-        let env = Environment::new();
-        build_lisp_function(&env, &document);
-        assert_eq!(do_lisp_env("(draw-string)", &env),"E1007");
-        assert_eq!(do_lisp_env("(draw-string \"Hello,World\" 20.0)", &env),"E1007");
-        assert_eq!(
-            do_lisp_env("(draw-string \"Hello,World\" 20.0 20.0 \"italic bold 20px sans-serif\" 10.0)", &env),
-            "E1007"
-        );
-        assert_eq!(
-            do_lisp_env("(draw-string 10.0 20.0 20.0 \"italic bold 20px sans-serif\")", &env),
-            "E1015"
-        );
-        assert_eq!(
-            do_lisp_env("(draw-string \"Hello,World\" 20.0 20.0 10.0)", &env),
-            "E1015"
-        );
-        assert_eq!(
-            do_lisp_env("(draw-string \"Hello,World\" #t 20.0 \"italic bold 20px sans-serif\")", &env),
-            "E1003"
-        );
-        assert_eq!(
-            do_lisp_env("(draw-string \"Hello,World\" 20.0 #t \"italic bold 20px sans-serif\")", &env),
-            "E1003"
-        );
-    }
-    #[wasm_bindgen_test]
-    fn draw_eval() {
-        let document = create_document();
-        let env = Environment::new();
-        build_lisp_function(&env, &document);
-
-        assert_eq!(do_lisp_env("(draw-eval)", &env), "E1007");
-        assert_eq!(do_lisp_env("(draw-eval (iota 20) 10)", &env), "E1007");
-    }
 }
